Refactor warrior objects into separate files
