+++ src/parse/DataScalar.ts
@@ -1,13 +1,13 @@
-import { GraphQLScalarType } from 'graphql';
+import { GraphQLScalarType } from 'graphql'
 
 export const DataScalar = new GraphQLScalarType({
     name: 'Data',
     description: 'Any data',
     parseValue(value: any) {
-        return value;
+        return value
     },
     serialize(value: any) {
-        return value;
+        return value
     },
     parseLiteral(ast) {
         return ast
