+++ src/warriors/MutationResult.ts
@@ -1,0 +1,10 @@
+import { Field } from 'type-graphql'
+
+export abstract class MutationResult<T> {
+    @Field()
+    success!: boolean
+    @Field({ nullable: true })
+    message?: string
+    @Field({ nullable: true })
+    result?: T
+}
