+++ src/warriors/WarriorResolver.ts
@@ -1,26 +1,12 @@
-import { Resolver, Mutation, Arg, ObjectType, Field, Query, Args } from "type-graphql";
-import { Warrior, WarriorInput } from "./Warrior";
-import Repository from "../database/Repository";
+import { Resolver, Mutation, ObjectType, Query, Args } from 'type-graphql'
+import { Warrior } from './Warrior'
+import Repository from '../database/Repository'
 import uuid from 'uuid/v1'
-import { corewar } from "corewar";
-import ParseResult from "../parse/ParseResult";
+import { corewar } from 'corewar'
+import { WarriorInput } from './WarriorInput'
+import { MutationResult } from './MutationResult'
 
 const WARRIOR_COLLECTION = 'warriors'
-
-abstract class MutationResult<T> {
-    @Field()
-    success!: boolean
-    @Field({ nullable: true })
-    message?: string
-    @Field({ nullable: true })
-    result?: T
-}
-
-@ObjectType()
-class SaveWarriorResult extends MutationResult<Warrior> { }
-
-@ObjectType()
-class DeleteWarriorResult extends MutationResult<string> { }
 
 class WarriorArgs {
     id!: string
@@ -30,19 +16,25 @@ class SaveWarriorArgs {
     warrior!: WarriorInput
 }
 
+@ObjectType()
+class SaveWarriorResult extends MutationResult<Warrior> { }
+
+@ObjectType()
+class DeleteWarriorResult extends MutationResult<string> { }
+
 @Resolver(Warrior)
 export class WarriorResolver {
-    @Query(returns => Warrior)
+    @Query(() => Warrior)
     async warrior(@Args() { id }: WarriorArgs): Promise<Warrior> {
         return (new Repository(WARRIOR_COLLECTION)).getById<Warrior>(id)
     }
 
-    @Query(returns => [Warrior])
+    @Query(() => [Warrior])
     async warriors(): Promise<Warrior[]> {
         return (new Repository(WARRIOR_COLLECTION)).getAll<Warrior>()
     }
 
-    @Mutation(returns => SaveWarriorResult)
+    @Mutation(() => SaveWarriorResult)
     async saveWarrior(@Args() { warrior }: SaveWarriorArgs): Promise<SaveWarriorResult> {
         let result = {
             ...warrior,
@@ -75,7 +67,7 @@ export class WarriorResolver {
         }
     }
 
-    @Mutation(returns => DeleteWarriorResult)
+    @Mutation(() => DeleteWarriorResult)
     async deleteWarrior(@Args() { id }: WarriorArgs): Promise<DeleteWarriorResult> {
         try {
             await (new Repository(WARRIOR_COLLECTION)).delete(id)
